refactor: consolidate contexts directory
