+++ src/components/RootClientProviders.tsx
@@ -1,7 +1,7 @@
 'use client'
 import React from 'react'
 import AuthSessionProvider from './AuthSessionProvider'
-import { SearchStatusProvider } from '@/context/SearchStatusContext'
+import { SearchStatusProvider } from '@/contexts/SearchStatusContext'
 
 export default function RootClientProviders({ children }: { children: React.ReactNode }) {
   return (
