+++ src/contexts/SearchStatusContext.tsx
